move the logic of updating ansible meta data to a separate location
